+++ src/tooling/piral-cli/src/apps/add-piral-instance-pilet.ts
@@ -4,7 +4,6 @@ import {
   progress,
   log,
   matchAnyPilet,
-  retrievePiletData,
   findFile,
   readJson,
   writeJson,